+++ include/malloy/rigid/rigid.hpp
@@ -2,6 +2,7 @@
 
 // Umbrella header for malloy_rigid: rigid bodies in two and three dimensions.
 #include <malloy/rigid/mass_properties.hpp>
+#include <malloy/rigid/mass_properties3d.hpp>
 #include <malloy/rigid/rigid_body2d.hpp>
 #include <malloy/rigid/rigid_body3d.hpp>
 #include <malloy/rigid/rigid3d_world.hpp>